Fix value calculation in eccentric connectivity index to retain float precision

+++ skfp/fingerprints/_new_mordred/descriptors/eccentric_connectivity_index.py
@@ -28,6 +28,6 @@ def calc(
     E = distance_matrix_regular.eccentricities
     D = adjacency_matrix_regular.degree
 
-    value = int((E * D).sum())
+    value = (E * D).sum()
 
     return np.array([value], dtype=np.float32), FEATURE_NAMES
